Crate ioctl_sys

Macros

io
Encode an ioctl command that has no associated data.

ioc
Encode an ioctl command.

ioctl
Declare a wrapper function around an ioctl.

ior
Encode an ioctl command that reads.

iorw
Encode an ioctl command that both reads and writes.

iow
Encode an ioctl command that writes.

Functions

ioc_dir
Extracts the "direction" (read/write/none) from an encoded ioctl command.

ioc_nr
Extracts the ioctl number from an encoded ioctl command.

ioc_size
Extracts the size from an encoded ioctl command.

ioc_type
Extracts the type from an encoded ioctl command.